+++ Test/Assets/Scripts/Component/CameraAreaManager.cs.meta
