+++ Test/Assets/Scripts/Enemy/Ability.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6b928037b4de384bbc7e82f6fc0fb07
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
