+++ Test/Assets/Scripts/Enemy/Basic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d819a85ea1bdec418b6e2ea46a23b7c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
